ai-assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
langchain==0.0.350
langchain-community==0.0.10
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pypdf==3.17.1
chromadb==0.4.18
sentence-transformers==2.2.2